refactor: getFileSize -> getFileSizes
