+++ src/app/(core)/teachers/programmes/[slug]/[tab]/Components/getFileSizes.ts
